pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "ebooklib>=0.20",
    "google-genai>=1.68.0",
    "python-dotenv>=1.2.2",
    "xhtml2pdf>=0.2.17",
]
